+++ app/src/main/java/com/example/plancton/navigation/router/EntryRouter.kt
@@ -1,8 +1,26 @@
 package com.example.plancton.navigation.router
 
+import com.example.plancton.navigation.screen.getLoginScreen
+import com.example.plancton.navigation.screen.getRegistrationScreen
+import com.github.terrakok.cicerone.Router
+import javax.inject.Inject
+
 interface EntryRouter {
 
     fun openLogin()
 
     fun openRegistration()
+}
+
+class EntryRouterImpl @Inject constructor(
+    private val router: Router,
+) : EntryRouter {
+
+    override fun openLogin() {
+        router.navigateTo(getLoginScreen())
+    }
+
+    override fun openRegistration() {
+        router.navigateTo(getRegistrationScreen())
+    }
 }
