+++ app/src/main/java/com/example/plancton/presentation/entry/EntryViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.plancton.presentation.viewmodel
+package com.example.plancton.presentation.entry
 
 import androidx.lifecycle.ViewModel
 import com.example.plancton.navigation.router.EntryRouter
