--- app/src/main/java/com/example/plancton/presentation/viewmodel/ViewModelFactory.kt
@@ -1,14 +1,0 @@
-package com.example.plancton.presentation.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-
-class ViewModelFactory @Inject constructor(
-    private val viewModelProviders: @JvmSuppressWildcards Map<Class<out ViewModel>, Provider<ViewModel>>,
-) : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>) =
-        viewModelProviders[modelClass]?.get() as T
-}
